tweak(server/classes/vehicle): don't error on invalid entity in initVehicle

+++ server/classes/vehicle.lua
@@ -582,10 +582,12 @@ AddStateBagChangeHandler("initVehicle", "", function(bagName, key, value, _, _)
     -- workaround for server-vehicles that exist in traffic randomly creating peds
     -- https://forum.cfx.re/t/sometimes-an-npc-spawns-inside-an-vehicle-spawned-with-createvehicleserversetter-or-create-automobile/4947251
     for i = -1, 0 do
-        local ped = GetPedInVehicleSeat(entity, i)
+        if DoesEntityExist(entity) then
+            local ped = GetPedInVehicleSeat(entity, i)
 
-        if not IsPedAPlayer(ped) then
-            DeleteEntity(ped)
+            if not IsPedAPlayer(ped) then
+                DeleteEntity(ped)
+            end
         end
     end
 end)
